Solves form resubmission problem

+++ src/org/care/presentation/seeker/DeleteJobServlet.java
@@ -1,6 +1,9 @@
 package org.care.presentation.seeker;
 
+import org.care.context.MyApplicationContext;
+import org.care.service.MemberService;
 import org.care.service.SeekerService;
+import org.care.utils.CommonUtil;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -12,19 +15,28 @@ import java.io.IOException;
 public class DeleteJobServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        int jobId = Integer.parseInt(req.getParameter("JobId"));
-        boolean isDeleted = SeekerService.deleteJob(jobId);
+        int userId = MyApplicationContext.get().getMember().getId();
 
-        if (isDeleted) {
-            req.setAttribute("msg", "Deleted Successfully!");
-        } else {
-            req.setAttribute("error", "Can't delete job");
+        String jobIdRaw = req.getParameter("JobId");
+        if (jobIdRaw != null && !jobIdRaw.isEmpty() && jobIdRaw.matches("^[0-9]+$")) {
+            int jobId = Integer.parseInt(jobIdRaw);
+
+            boolean isDeleted = false;
+            if(userId == SeekerService.getUserIdforJobId(jobId)) {
+                isDeleted = SeekerService.deleteJob(jobId);
+            }
+
+            if (isDeleted) {
+                resp.sendRedirect(CommonUtil.getRedirectURL("/seeker/list-job?success=true"));
+            } else {
+                resp.sendRedirect(CommonUtil.getRedirectURL("/seeker/list-job?success=false"));
+            }
         }
-        resp.sendRedirect("/HomeJobMarketplace/seeker/list-job");
+
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-
+        resp.sendRedirect(CommonUtil.getRedirectURL(""));
     }
 }
